fix: refactor package name and improve server initialization in tests
